feat: added option to keep directions in graphs
With this addition, a distinct separation between
directed and undirected graphs is made. The user
can keep the directions in road/network graphs or
allow the engine transform to undirected graphs.
Because of the road/network analysis in the engine,
by default the graphs are converted to undirected
graphs. This option can be manipulated via settings page.