fix: a minor bug implementing facility access
The metrics (1,4 and 7) which use facility accessibility in the calculation has been using
'facility_access' attribute in the individuals table. Depending on the road
connectivity analysis, this metric is correcly assigned True of False. However
if an individual is not associated with a facility, then this attribute is also set to false.
For this reason, the population displacement results is slightly larger than expected.
This minor issue has no effect on metric 1 and 4 since only the workers and students
are used in the calculation of that metrics.