feat: redefine population displacement (metric7)
An individual is assumed to be displaced when any of the following conditions hold:

At least X of the following assets either suffer direct damage, cannot be reached via the
transportation network (applies to all except residential buildings), or have no electricity:

 - the individual's residence, workplace, school, or associated hospital

where X is the population displacement consensus which can be adjusted in the settings with a default value of 2.

With this redefinition, metric7 (population displacement) is much lower than when
calculated with the old definition. (Old definition is simply the new one with consensis=1)