fix: corrected attribute list of power node
The "eq_vuln" attribute is replaced with "eq_frgl".
In the first webapp development, the taxonomy string was defined by means of "eq_vuln". Later, it is changed to
"eq_frgl". However it is forgotten in the application state dictionary. This mistake has not caused any errors
in calculation. Neverthless, we change it with the correct one which is "eq_frgl".