fix: center calculation is now in geometric
As soon as a geo dataframe is uploaded, its center is calculated so that
the data will be centered on the map. It has been calculated in geographic
coordinated which produces warnings. To get rid of this,
geodataframe is converted first to geometry, then calculation takes place,
and finally converted back to geography.