ui: give extra info after data generation
When the data generation is completed, the ui gives
an extra information stating that the building, household
and individual are created and the user can upload hazard
and vulnerability/fragility data.

The default form of the message is added to the .env.global
file via data_generation_complete_message keyword. The message can be
customized by setting data_generation_complete_message variable in
.env.global or .env.local file. If the key is not defined
the hard-coded string in the code is used.